feat: source example (#34)
