+++ apps/example/src/app.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import bodyRouter from './routers/body'
+import sourceRouter from './routers/source'
 
 const app = express()
 
@@ -8,6 +8,6 @@ const port = process.env.PORT || 3000
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
 
-app.use(bodyRouter)
+app.use(sourceRouter)
 
 app.listen(port, () => {console.log(`Example app listening on port ${port}`)})
